pyproject.toml
[project]
name = "tod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.9.0",
    "pydantic>=2.11.7",
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "aci-sdk>=1.0.0b3",
]
